Replaced player graphic with a 3d animated model
Note the model is not source controlled. You must get it and add it to
the project yourself.
https://www.unrealengine.com/marketplace/en-US/slug/paragon-gideon
