+++ Config/DefaultEditor.ini
@@ -6,3 +6,5 @@ bAllowClassAndBlueprintPinMatching=true
 bReplaceBlueprintWithClass= true
 bDontLoadBlueprintOutsideEditor= true
 bBlueprintIsNotBlueprintType= true
+
+
