+++ Source/CS330_Final/CS330_FinalPawn.cpp
@@ -23,7 +23,7 @@ const FName ACS330_FinalPawn::FireBinding("Fire");
 ACS330_FinalPawn::ACS330_FinalPawn()
 {	
 	//set bullet
-	static ConstructorHelpers::FClassFinder<ACS330_FinalProjectile> BulletBPClass(TEXT("/Game/MyCS330_FinalProjectile"));
+	static ConstructorHelpers::FClassFinder<ACS330_FinalProjectile> BulletBPClass(TEXT("/Game/Blueprints/Bullets/BP_TestProjectile"));
 	if (BulletBPClass.Class != NULL)
 	{
 		Bullet = BulletBPClass.Class;
