+++ Source/CS330_Final/PlayerCharacter.cpp
@@ -1,0 +1,182 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "PlayerCharacter.h"
+#include "CS330_FinalProjectile.h"
+#include "TimerManager.h"
+#include "UObject/ConstructorHelpers.h"
+#include "Camera/CameraComponent.h"
+#include "Components/StaticMeshComponent.h"
+#include "Components/InputComponent.h"
+#include "GameFramework/SpringArmComponent.h"
+#include "Engine/CollisionProfile.h"
+#include "Engine/StaticMesh.h"
+#include "Kismet/GameplayStatics.h"
+#include "Kismet/KismetMathLibrary.h"
+#include "Sound/SoundBase.h"
+
+const FName APlayerCharacter::MoveForwardBinding("MoveForward");
+const FName APlayerCharacter::MoveRightBinding("MoveRight");
+const FName APlayerCharacter::FireForwardBinding("FireForward");
+const FName APlayerCharacter::FireRightBinding("FireRight");
+const FName APlayerCharacter::FireBinding("Fire");
+
+APlayerCharacter::APlayerCharacter()
+{
+	//set bullet
+	static ConstructorHelpers::FClassFinder<ACS330_FinalProjectile> BulletBPClass(TEXT("/Game/Blueprints/Bullets/BP_TestProjectile"));
+	if (BulletBPClass.Class != NULL)
+	{
+		Bullet = BulletBPClass.Class;
+	}
+
+	// Cache our sound effect
+	static ConstructorHelpers::FObjectFinder<USoundBase> FireAudio(TEXT("/Game/TwinStick/Audio/TwinStickFire.TwinStickFire"));
+	FireSound = FireAudio.Object;
+
+	// Create a camera boom...
+	CameraBoom = CreateDefaultSubobject<USpringArmComponent>(TEXT("CameraBoom"));
+	CameraBoom->SetupAttachment(RootComponent);
+	CameraBoom->bAbsoluteRotation = true; // Don't want arm to rotate when ship does
+	CameraBoom->TargetArmLength = 1200.f;
+	CameraBoom->RelativeRotation = FRotator(-80.f, 0.f, 0.f);
+	CameraBoom->bDoCollisionTest = false; // Don't want to pull camera in when it collides with level
+
+	// Create a camera...
+	CameraComponent = CreateDefaultSubobject<UCameraComponent>(TEXT("TopDownCamera"));
+	CameraComponent->SetupAttachment(CameraBoom, USpringArmComponent::SocketName);
+	CameraComponent->bUsePawnControlRotation = false;	// Camera does not rotate relative to arm
+
+	// Movement
+	MoveSpeed = 1000.0f;
+	// Weapon
+	GunOffset = FVector(90.f, 0.f, 0.f);
+	FireRate = 0.1f;
+	bCanFire = true;
+
+	APlayerController* PC = Cast<APlayerController>(GetController());
+
+	if (PC)
+	{
+		PC->bShowMouseCursor = true;
+		PC->bEnableClickEvents = true;
+		PC->bEnableMouseOverEvents = true;
+	}
+
+}
+
+void APlayerCharacter::SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent)
+{
+	check(PlayerInputComponent);
+
+	// set up gameplay key bindings
+	PlayerInputComponent->BindAxis(MoveForwardBinding);
+	PlayerInputComponent->BindAxis(MoveRightBinding);
+	PlayerInputComponent->BindAxis(FireForwardBinding);
+	PlayerInputComponent->BindAxis(FireRightBinding);
+	PlayerInputComponent->BindAxis(FireBinding);
+}
+
+void APlayerCharacter::Tick(float DeltaSeconds)
+{
+	// Find movement direction
+	const float ForwardValue = GetInputAxisValue(MoveForwardBinding);
+	const float RightValue = GetInputAxisValue(MoveRightBinding);
+	// Clamp max size so that (X=1, Y=1) doesn't cause faster movement in diagonal directions
+	const FVector MoveDirection = FVector(ForwardValue, RightValue, 0.f).GetClampedToMaxSize(1.0f);
+
+	// Calculate  movement
+	const FVector Movement = MoveDirection * MoveSpeed * DeltaSeconds;
+
+	// If non-zero size, move this actor
+	if (Movement.SizeSquared() > 0.0f)
+	{
+		const FRotator NewRotation = Movement.Rotation();
+		FHitResult Hit(1.f);
+		RootComponent->MoveComponent(Movement, NewRotation, true, &Hit);
+
+		if (Hit.IsValidBlockingHit())
+		{
+			const FVector Normal2D = Hit.Normal.GetSafeNormal2D();
+			const FVector Deflection = FVector::VectorPlaneProject(Movement, Normal2D) * (1.f - Hit.Time);
+			RootComponent->MoveComponent(Deflection, NewRotation, true);
+		}
+	}
+
+	// Create fire direction vector
+	const float FireForwardValue = GetInputAxisValue(FireForwardBinding);
+	const float FireRightValue = GetInputAxisValue(FireRightBinding);
+	const FVector FireDirection = GetActorForwardVector();
+
+	// Try and fire a shot
+	UpdateMouseLook();
+	bool Shoot = GetInputAxisValue(FireBinding);
+	if (Shoot)
+		FireShot();
+}
+
+void APlayerCharacter::UpdateMouseLook()
+{
+	UWorld* World = GetWorld();
+	if (World)
+	{
+		APlayerController* PlayerController = UGameplayStatics::GetPlayerController(GetWorld(), 0);
+		if (PlayerController)
+		{
+			FVector MouseLocation;
+			FVector MouseDirection;
+			PlayerController->DeprojectMousePositionToWorld(MouseLocation, MouseDirection);
+
+			MouseDirection *= 100000;
+			//MouseDirection
+			FHitResult hitResult;
+			bool hit;
+			hit = World->LineTraceSingleByChannel(hitResult, MouseLocation, MouseLocation + MouseDirection, ECollisionChannel::ECC_Visibility);
+			if (hit)
+			{
+				FVector targetLocation = hitResult.Location;
+				FRotator PlayerRot = UKismetMathLibrary::FindLookAtRotation(GetActorLocation(), targetLocation);
+				this->SetActorRotation(FRotator(0.0f, PlayerRot.Yaw, 0.f));
+			}
+
+		}
+	}
+}
+void APlayerCharacter::FireShot()
+{
+	FVector FireDirection = GetActorForwardVector();
+	// If it's ok to fire again
+	if (bCanFire == true)
+	{
+		// If we are pressing fire stick in a direction
+		if (FireDirection.SizeSquared() > 0.0f)
+		{
+			const FRotator FireRotation = FireDirection.Rotation();
+			// Spawn projectile at an offset from this pawn
+			const FVector SpawnLocation = GetActorLocation() + FireRotation.RotateVector(GunOffset);
+
+			UWorld* const World = GetWorld();
+			if (World != NULL)
+			{
+				// spawn the projectile
+				World->SpawnActor<ACS330_FinalProjectile>(Bullet, SpawnLocation, FireRotation);
+			}
+
+			bCanFire = false;
+			World->GetTimerManager().SetTimer(TimerHandle_ShotTimerExpired, this, &APlayerCharacter::ShotTimerExpired, FireRate);
+
+			// try and play the sound if specified
+			if (FireSound != nullptr)
+			{
+				UGameplayStatics::PlaySoundAtLocation(this, FireSound, GetActorLocation());
+			}
+
+			bCanFire = false;
+		}
+	}
+}
+
+void APlayerCharacter::ShotTimerExpired()
+{
+	bCanFire = true;
+}
+
