--- enc_temp_folder/7fb9fcb5adeb3aa665a25cc08b1fa4/CS330_FinalPawn.cpp
@@ -1,182 +1,0 @@
-// Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
-
-#include "CS330_FinalPawn.h"
-#include "CS330_FinalProjectile.h"
-#include "TimerManager.h"
-#include "UObject/ConstructorHelpers.h"
-#include "Camera/CameraComponent.h"
-#include "Components/StaticMeshComponent.h"
-#include "Components/InputComponent.h"
-#include "GameFramework/SpringArmComponent.h"
-#include "Engine/CollisionProfile.h"
-#include "Engine/StaticMesh.h"
-#include "Kismet/GameplayStatics.h"
-#include "Kismet/KismetMathLibrary.h"
-#include "Sound/SoundBase.h"
-
-const FName ACS330_FinalPawn::MoveForwardBinding("MoveForward");
-const FName ACS330_FinalPawn::MoveRightBinding("MoveRight");
-const FName ACS330_FinalPawn::FireForwardBinding("FireForward");
-const FName ACS330_FinalPawn::FireRightBinding("FireRight");
-const FName ACS330_FinalPawn::FireBinding("Fire");
-
-ACS330_FinalPawn::ACS330_FinalPawn()
-{	
-	static ConstructorHelpers::FObjectFinder<UStaticMesh> ShipMesh(TEXT("/Game/TwinStick/Meshes/TwinStickUFO.TwinStickUFO"));
-	// Create the mesh component
-	ShipMeshComponent = CreateDefaultSubobject<UStaticMeshComponent>(TEXT("ShipMesh"));
-	RootComponent = ShipMeshComponent;
-	ShipMeshComponent->SetCollisionProfileName(UCollisionProfile::Pawn_ProfileName);
-	ShipMeshComponent->SetStaticMesh(ShipMesh.Object);
-	
-	// Cache our sound effect
-	static ConstructorHelpers::FObjectFinder<USoundBase> FireAudio(TEXT("/Game/TwinStick/Audio/TwinStickFire.TwinStickFire"));
-	FireSound = FireAudio.Object;
-
-	// Create a camera boom...
-	CameraBoom = CreateDefaultSubobject<USpringArmComponent>(TEXT("CameraBoom"));
-	CameraBoom->SetupAttachment(RootComponent);
-	CameraBoom->bAbsoluteRotation = true; // Don't want arm to rotate when ship does
-	CameraBoom->TargetArmLength = 1200.f;
-	CameraBoom->RelativeRotation = FRotator(-80.f, 0.f, 0.f);
-	CameraBoom->bDoCollisionTest = false; // Don't want to pull camera in when it collides with level
-
-	// Create a camera...
-	CameraComponent = CreateDefaultSubobject<UCameraComponent>(TEXT("TopDownCamera"));
-	CameraComponent->SetupAttachment(CameraBoom, USpringArmComponent::SocketName);
-	CameraComponent->bUsePawnControlRotation = false;	// Camera does not rotate relative to arm
-
-	// Movement
-	MoveSpeed = 1000.0f;
-	// Weapon
-	GunOffset = FVector(90.f, 0.f, 0.f);
-	FireRate = 0.1f;
-	bCanFire = true;
-
-	APlayerController* PC = Cast<APlayerController>(GetController());
-
-	if (PC)
-	{
-		PC->bShowMouseCursor = true;
-		PC->bEnableClickEvents = true;
-		PC->bEnableMouseOverEvents = true;
-	}
-
-}
-
-void ACS330_FinalPawn::SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent)
-{
-	check(PlayerInputComponent);
-
-	// set up gameplay key bindings
-	PlayerInputComponent->BindAxis(MoveForwardBinding);
-	PlayerInputComponent->BindAxis(MoveRightBinding);
-	PlayerInputComponent->BindAxis(FireForwardBinding);
-	PlayerInputComponent->BindAxis(FireRightBinding);
-	PlayerInputComponent->BindAxis(FireBinding);
-}
-
-void ACS330_FinalPawn::Tick(float DeltaSeconds)
-{
-	// Find movement direction
-	const float ForwardValue = GetInputAxisValue(MoveForwardBinding);
-	const float RightValue = GetInputAxisValue(MoveRightBinding);
-	// Clamp max size so that (X=1, Y=1) doesn't cause faster movement in diagonal directions
-	const FVector MoveDirection = FVector(ForwardValue, RightValue, 0.f).GetClampedToMaxSize(1.0f);
-
-	// Calculate  movement
-	const FVector Movement = MoveDirection * MoveSpeed * DeltaSeconds;
-
-	// If non-zero size, move this actor
-	if (Movement.SizeSquared() > 0.0f)
-	{
-		const FRotator NewRotation = Movement.Rotation();
-		FHitResult Hit(1.f);
-		RootComponent->MoveComponent(Movement, NewRotation, true, &Hit);
-		
-		if (Hit.IsValidBlockingHit())
-		{
-			const FVector Normal2D = Hit.Normal.GetSafeNormal2D();
-			const FVector Deflection = FVector::VectorPlaneProject(Movement, Normal2D) * (1.f - Hit.Time);
-			RootComponent->MoveComponent(Deflection, NewRotation, true);
-		}
-	}
-	
-	// Create fire direction vector
-	const float FireForwardValue = GetInputAxisValue(FireForwardBinding);
-	const float FireRightValue = GetInputAxisValue(FireRightBinding);
-	const FVector FireDirection = GetActorForwardVector();
-
-	// Try and fire a shot
-	UpdateMouseLook();
-	bool Shoot = GetInputAxisValue(FireBinding);
-	if(Shoot)
-		FireShot();
-}
-
-void ACS330_FinalPawn::UpdateMouseLook()
-{
-	UWorld* World = GetWorld();
-	if (World)
-	{
-		APlayerController* PlayerController = UGameplayStatics::GetPlayerController(GetWorld(), 0);
-		if (PlayerController)
-		{
-			FVector MouseLocation;
-			FVector MouseDirection;
-			PlayerController->DeprojectMousePositionToWorld(MouseLocation, MouseDirection);
-
-			MouseDirection *= 100000;
-			//MouseDirection
-			FHitResult hitResult;
-			bool hit;
-			hit = World->LineTraceSingleByChannel(hitResult, MouseLocation, MouseLocation+MouseDirection, ECollisionChannel::ECC_Visibility);
-			if (hit)
-			{
-				FVector targetLocation = hitResult.Location;
-				FRotator PlayerRot = UKismetMathLibrary::FindLookAtRotation(ShipMeshComponent->GetComponentLocation(), targetLocation);
-				ShipMeshComponent->SetWorldRotation(FRotator(0.0f,PlayerRot.Yaw,0.f));
-			}
-
-		}
-	}
-}
-void ACS330_FinalPawn::FireShot()
-{
-	FVector FireDirection = GetActorForwardVector();
-	// If it's ok to fire again
-	if (bCanFire == true)
-	{
-		// If we are pressing fire stick in a direction
-		if (FireDirection.SizeSquared() > 0.0f)
-		{
-			const FRotator FireRotation = FireDirection.Rotation();
-			// Spawn projectile at an offset from this pawn
-			const FVector SpawnLocation = GetActorLocation() + FireRotation.RotateVector(GunOffset);
-
-			UWorld* const World = GetWorld();
-			if (World != NULL)
-			{
-				// spawn the projectile
-				World->SpawnActor<ACS330_FinalProjectile>(SpawnLocation, FireRotation);
-			}
-
-			bCanFire = false;
-			World->GetTimerManager().SetTimer(TimerHandle_ShotTimerExpired, this, &ACS330_FinalPawn::ShotTimerExpired, FireRate);
-
-			// try and play the sound if specified
-			if (FireSound != nullptr)
-			{
-				UGameplayStatics::PlaySoundAtLocation(this, FireSound, GetActorLocation());
-			}
-
-			bCanFire = false;
-		}
-	}
-}
-
-void ACS330_FinalPawn::ShotTimerExpired()
-{
-	bCanFire = true;
-}
-
